# specs/001-artifact-registry/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Indiana Jones Artifact Registry API
  description: |
    Open-source artifact registry API supporting multiple package formats.
    This is the administrative and generic artifact API. Format-specific
    endpoints (Maven, npm, Docker, etc.) are defined in separate specs.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /repositories:
    get:
      tags: [repositories]
      summary: List repositories
      operationId: listRepositories
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: format
          in: query
          schema:
            $ref: '#/components/schemas/RepositoryFormat'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/RepositoryType'
      responses:
        '200':
          description: List of repositories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryList'

    post:
      tags: [repositories]
      summary: Create repository
      operationId: createRepository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRepositoryRequest'
      responses:
        '201':
          description: Repository created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /repositories/{key}:
    parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
        description: Repository key

    get:
      tags: [repositories]
      summary: Get repository details
      operationId: getRepository
      responses:
        '200':
          description: Repository details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [repositories]
      summary: Update repository
      operationId: updateRepository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRepositoryRequest'
      responses:
        '200':
          description: Repository updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [repositories]
      summary: Delete repository
      operationId: deleteRepository
      parameters:
        - name: force
          in: query
          schema:
            type: boolean
            default: false
          description: Force delete even if repository contains artifacts
      responses:
        '204':
          description: Repository deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  # ============================================================================
  # Artifacts (Generic API)
  # ============================================================================
  /repositories/{key}/artifacts:
    parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string

    get:
      tags: [artifacts]
      summary: List artifacts in repository
      operationId: listArtifacts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: q
          in: query
          schema:
            type: string
          description: Search query
        - name: path_prefix
          in: query
          schema:
            type: string
          description: Filter by path prefix
      responses:
        '200':
          description: List of artifacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactList'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    # Health
...
    RepositoryFormat:
      type: string
      enum:
        - maven
        - gradle
        - npm
        - pypi
        - nuget
        - go
        - rubygems
        - docker
        - helm
        - rpm
        - debian
        - conan
        - cargo
        - generic
    RepositoryType:
      type: string
      enum: [local, remote, virtual]
    Repository:
      type: object
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        name:
          type: string
        description:
          type: string
        format:
          $ref: '#/components/schemas/RepositoryFormat'
        repo_type:
          $ref: '#/components/schemas/RepositoryType'
        is_public:
          type: boolean
        artifact_count:
          type: integer
        storage_used_bytes:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RepositoryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Repository'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateRepositoryRequest:
      type: object
      required: [key, name, format, repo_type]
      properties:
        key:
          type: string
          pattern: '^[a-z0-9][a-z0-9-]*[a-z0-9]$'
          minLength: 3
          maxLength: 255
        name:
          type: string
        description:
          type: string
        format:
          $ref: '#/components/schemas/RepositoryFormat'
        repo_type:
          $ref: '#/components/schemas/RepositoryType'
        is_public:
          type: boolean
          default: false
        upstream_url:
          type: string
          format: uri
          description: Required for remote repositories
    UpdateRepositoryRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        is_public:
          type: boolean

    # Artifact
    Artifact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        repository_key:
          type: string
        path:
          type: string
        name:
          type: string
        version:
          type: string
        size_bytes:
          type: integer
          format: int64
        checksum_sha256:
          type: string
        content_type:
          type: string
        download_count:
          type: integer
        created_at:
          type: string
          format: date-time
        metadata:
          type: object
          description: Format-specific metadata
    ArtifactList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Artifact'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
security:
  - bearerAuth: []